+++ be-engine-v1/be-other-app-wx/src/main/java/com/cmcorg20230301/be/engine/other/app/wx/model/vo/WxAccessTokenVO.java
@@ -15,6 +15,6 @@ public class WxAccessTokenVO extends WxBaseVO {
 
     @Alias(value = "expires_in")
     @Schema(description = "凭证有效时间，单位：秒。目前是7200秒之内的值。")
-    private Integer expiresIn;
+    private Long expiresIn;
 
 }
