+++ be-engine-v1/be-other-app-wx/src/main/java/com/cmcorg20230301/be/engine/other/app/wx/model/vo/WxOpenIdVO.java
@@ -21,6 +21,6 @@ public class WxOpenIdVO extends WxBaseVO {
 
     @Alias(value = "expires_in")
     @Schema(description = "access_token接口调用凭证超时时间，单位（秒）")
-    private Integer expiresIn;
+    private Long expiresIn;
 
 }
