+++ be-engine-v1/be-pay-google/src/main/java/com/cmcorg20230301/be/engine/pay/google/util/PayGoogleUtil.java
@@ -165,7 +165,7 @@ public class PayGoogleUtil {
 
         }
 
-        Integer expiresIn = jsonObject.getInt("expires_in"); // 这里的单位是：秒
+        Long expiresIn = jsonObject.getLong("expires_in"); // 这里的单位是：秒
 
         CacheRedisKafkaLocalUtil
             .put(BaseRedisKeyEnum.GOOGLE_ACCESS_TOKEN_CACHE, sufKey, null, expiresIn * 1000, () -> accessTokenResult);
